Use a configurable timeout when performing the connectivity check
The connectivity check to determine whether the cluster subnets are
proxied that the root daemon performs when it boots up, was hard-coded
to 2 seconds, which meant that the normal case (when the cluster isn't
proxied), would always block the network initialization that long.

This commit shortens the timeout 500 ms, and makes it configurable. The
check can be disabled by setting the timeout to zero.

Signed-off-by: Thomas Hallgren <thomas@datawire.io>